+++ src/main/java/com/peaksoft/dto/UserResponse.java
@@ -1,15 +1,10 @@
 package com.peaksoft.dto;
 
-import com.peaksoft.model.entity.enums.Country;
-import com.peaksoft.model.entity.enums.Role;
+
+import com.peaksoft.model.entity.enums.RoleE;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-
-import javax.persistence.Column;
-import javax.validation.constraints.Email;
-import java.time.LocalDate;
 
 @Getter @Setter
 @Builder
@@ -20,6 +15,6 @@ public class UserResponse  {
     private String lastName;
     private String email;
     private String password;
-    private Role role;
+    private RoleE role;
 
 }
