+++ src/main/java/com/peaksoft/repository/UserRepository.java
@@ -2,8 +2,11 @@ package com.peaksoft.repository;
 
 import com.peaksoft.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
 
+@Repository
 public interface UserRepository extends JpaRepository<User,Long> {
-    User findByEmail(String email);
+    Optional<User> findByEmail(String email);
 }
