+++ src/main/java/com/peaksoft/servise/UserServiceImpl.java
@@ -1,37 +1,25 @@
 package com.peaksoft.servise;
 
-import com.peaksoft.model.User;
-import com.peaksoft.model.entity.enums.Role;
+
 import com.peaksoft.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-import org.webjars.NotFoundException;
+import org.springframework.stereotype.Component;
 
-import javax.transaction.Transactional;
-import java.util.Collection;
-import java.util.Optional;
-import java.util.stream.Collectors;
 
-import static java.lang.String.format;
-
-@Service
-@RequiredArgsConstructor
+@Component
 public class UserServiceImpl implements UserDetailsService {
     private final UserRepository userRepository;
 
+
+    public UserServiceImpl(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
     @Override
-    @Transactional
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        User user = userRepository.findByEmail(username);
-        if (user != null) {
-            return user;
-        } else {
-            throw new UsernameNotFoundException(format("User with email - %s, not found", username));
-        }
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+        return userRepository.findByEmail(email).orElseThrow(()
+                ->  new UsernameNotFoundException(email.format("User not found")));
     }
 }
