--- src/main/java/com/peaksoft/spring_boot_jwt_token/dto/UserRequest.java
@@ -1,13 +1,0 @@
-package com.peaksoft.spring_boot_jwt_token.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter @Setter
-public class UserRequest {
-
-    private String firstName;
-    private String lastName;
-    private String email;
-    private String password;
-}
