--- src/main/java/com/peaksoft/spring_boot_jwt_token/dto/UserResponse.java
@@ -1,18 +1,0 @@
-package com.peaksoft.spring_boot_jwt_token.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Builder
-public class UserResponse {
-
-    private Long id;
-    private String firstName;
-    private String lastName;
-    private String email;
-    private String password;
-
-}
