--- src/main/java/com/peaksoft/spring_boot_jwt_token/dto/ValidationType.java
@@ -1,8 +1,0 @@
-package com.peaksoft.spring_boot_jwt_token.dto;
-
-public class ValidationType {
-
-    public static final String SUCCESSFUL = "successful";
-    public static final String LOGIN_MAPPER = "mapper";
-
-}
