--- src/main/java/com/peaksoft/spring_boot_jwt_token/repository/UserRepository.java
@@ -1,14 +1,0 @@
-package com.peaksoft.spring_boot_jwt_token.repository;
-
-
-import com.peaksoft.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository <User, Long > {
-
-        Optional<User> findByEmail(String email);
-}
