--- src/main/java/com/peaksoft/spring_boot_jwt_token/service/UserServiceImpl.java
@@ -1,26 +1,0 @@
-package com.peaksoft.spring_boot_jwt_token.service;
-
-import com.peaksoft.spring_boot_jwt_token.repository.UserRepository;
-
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Component;
-
-
-@Component
-public class UserServiceImpl implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    public UserServiceImpl(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        return userRepository.findByEmail(email).orElseThrow(()
-                ->  new UsernameNotFoundException(email.format("User not found")));
-    }
-
-}
